完善商品修改添加功能；开发商品详情功能；

+++ member_management/src/main/java/com/qiqi/member_management/management/business/model/ProductInfo.java
@@ -1,5 +1,6 @@
 package com.qiqi.member_management.management.business.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class ProductInfo {
@@ -13,7 +14,7 @@ public class ProductInfo {
 
     private String storageId;
 
-    private Long price;
+    private BigDecimal price;
 
     private String describute;
 
@@ -73,11 +74,11 @@ public class ProductInfo {
         this.storageId = storageId == null ? null : storageId.trim();
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
